requirements.txt
Django==3.1.7
django-shopping-cart==0.1
Pillow==8.1.2
django-crispy-forms
